+++ config/lstm.yml
@@ -13,6 +13,10 @@ model:
     - num_units: 200
     - num_units: 200
   loss_func: sequence_loss
-
-
-
+train:
+  epoch_num: 10
+  num_steps: 20
+  batch_size: 20
+  keep_prob: 1.0
+  gradient_clip: global_norm
+  optimizer: GradientDescent
