+++ test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 source ~/tensorflow-gpu/bin/activate
-nohup `python -m py.ptb_word_test` > test_out.log &
+nohup `python -m py.test_lstm` > test_out.log &
